Use JVM's CLDR data for locales and work around differences
To achieve this, we need to fork a JVM in our build, so that we can
specify JVM options.
We also need to set our base directory, otherwise file-based tests look
in the wrong folder.

Now that we finally managed to use the JVM's CLDR data, we have to deal
with the differences between JRE locale data and CLDR data. Example:
JRE specifies capitalized brazilian month and day-of-week names,
while CLDR specifies lower-cased names.